feat(dashboard): qa_score en escala 0-10 y badge en FlowView
- instrumentation: QAScorePayload.Score/MaxScore pasan de int a float64
- store: nuevo metodo GetRunSummary para leer un run por ID
- app: binding Wails GetRunSummary expuesto al frontend
- frontend: componente QABadge reutilizable (reemplaza QACell inline)
- flow-view: muestra el QA badge en el header cuando hay qa_score
- formatQAScore: umbrales 0-10 (>=8 success, >=6 warn, <6 fail) con sufijo "/10"

+++ internal/instrumentation/events.go
@@ -95,11 +95,11 @@ type FileTouchedPayload struct {
 }
 
 type QAScorePayload struct {
-	AgentID  string         `json:"agent_id"`
-	Score    int            `json:"score"`
-	MaxScore int            `json:"max_score"`
-	Criteria map[string]int `json:"criteria"`
-	Notes    string         `json:"notes,omitempty"`
+	AgentID  string             `json:"agent_id"`
+	Score    float64            `json:"score"`
+	MaxScore float64            `json:"max_score"`
+	Criteria map[string]float64 `json:"criteria"`
+	Notes    string             `json:"notes,omitempty"`
 }
 
 // NewEvent creates a new Event with a UUID v4 event ID, UTC timestamp, and
